requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.0.0
langgraph>=0.2.0
langchain-openai>=0.1.0
langchain-core>=0.2.0
openai>=1.12.0
python-dotenv>=1.0.0

